backend/requirements.txt
fastapi
uvicorn
openai
google-generativeai
qdrant-client
python-dotenv
pydantic
psycopg[binary]
sqlalchemy
alembic
